+++ extension/src/util/FileWatcher.ts
