--- extension/src/watcher/index.ts
@@ -1,1 +1,0 @@
-export * from './FileWatcher';
